hibernate version updated with second level cache implementation

+++ idas.chox.data/src/main/java/idas/chox/data/services/BordereauServiceImpl.java
@@ -118,7 +118,7 @@ public class BordereauServiceImpl extends SecureDataService implements Bordereau
         criteria.setProjection(Projections.rowCount());
         List totalCountResult = criteria.list();
         criteria.setProjection(null);
-        return (Integer) totalCountResult.get(0);
+        return ((Long) totalCountResult.get(0)).intValue();
     }
